feat: add `App` component
