+++ src/index.tsx
@@ -1,15 +1,13 @@
-// Styles
 import "./styles/app.css";
 
 import { App } from "app";
-// React
 import React from "react";
 import ReactDOM from "react-dom";
-import { HashRouter } from "react-router-dom";
+import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.render(
-	<HashRouter>
+	<BrowserRouter>
 		<App />
-	</HashRouter>,
+	</BrowserRouter>,
 	document.getElementById("root"),
 );
